confirmation sweet alert implemented

+++ src/Components/MyCart/MyCart.jsx
@@ -4,6 +4,7 @@ import { AuthContext } from "../../ContextProvider/ContextProvider";
 // import MyCartDetails from "../MyCartDetails/MyCartDetails";
 import { FaStar } from 'react-icons/fa';
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 
 const MyCart = () => {
@@ -19,18 +20,43 @@ const MyCart = () => {
     }, [user?.email]);
 
     const deleteHandle = (id) => {
-        fetch(`http://localhost:5000/userCart/${id}`, {
-            method: 'DELETE'
-        })
-        .then(res=>res.json())
-        .then(data=>{
-            console.log(data)
-            if(data.deletedCount){
-                alert('user is deleted ')
-                const remaining = addCarts.filter(cart=>cart._id !== id);
-                setAddCarts(remaining)
+
+
+
+        Swal.fire({
+            title: 'Are you sure to delete?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+
+                fetch(`http://localhost:5000/userCart/${id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data)
+                        if (data.deletedCount) {
+                            Swal.fire(
+                                'Deleted!',
+                                'Your cart product  has been deleted.',
+                                'success'
+                            )
+                            const remaining = addCarts.filter(cart => cart._id !== id);
+                            setAddCarts(remaining)
+                        }
+                    })
+
             }
         })
+
+
+
+
 
     }
 
@@ -40,7 +66,7 @@ const MyCart = () => {
         <div>
             <h3>This is the My cart component:{addCarts.length} </h3>
             <div className="grid md:grid-cols-2 lg:grid-cols-3 gap-4">
-                
+
                 {
                     addCarts.map((addCart, index) => <p key={index}>
